Restructure everything! Also add Dockerfile for easier and more consistent running
