+++ Dockerfile
@@ -1,0 +1,30 @@
+FROM python:3.8.6-slim-buster as base
+
+ENV PYTHONFAULTHANDLER=1 \
+    PYTHONHASHSEED=random \
+    PYTHONUNBUFFERED=1
+
+WORKDIR /app
+
+FROM base as builder
+
+ENV VIRTUAL_ENV=/venv \
+    PIP_DEFAULT_TIMEOUT=100 \
+    PIP_DISABLE_PIP_VERSION_CHECK=1 \
+    PIP_NO_CACHE_DIR=1 \
+    POETRY_VERSION=1.1.4
+
+RUN pip install "poetry==$POETRY_VERSION"
+RUN python -m venv $VIRTUAL_ENV
+ENV PATH="$VIRTUAL_ENV/bin:$PATH"
+
+COPY pyproject.toml poetry.lock ./
+RUN poetry export -f requirements.txt | pip install -r /dev/stdin
+
+COPY . .
+RUN poetry build && pip install dist/*.whl
+
+FROM base as final
+
+COPY --from=builder /venv /venv
+CMD ["python", "nerdlandbot"]
