+++ nerdlandbot/__main__.py
@@ -1,0 +1,57 @@
+import os
+import sys
+import discord
+
+from dotenv import load_dotenv
+
+from nerdlandbot.bot import NerdlandBot
+from nerdlandbot.helpers.log import info, fatal
+from nerdlandbot.translations.Translations import get_text as _
+from nerdlandbot.scheduler.Scheduler import check_and_post_latest_videos
+
+load_dotenv()
+
+PREFIX = os.getenv("PREFIX")
+TOKEN = os.getenv("DISCORD_TOKEN")
+
+if PREFIX:
+    info("Start bot with prefix '" + PREFIX + "'")
+else:
+    fatal("Please provide a PREFIX in your .env file")
+    sys.exit()
+
+# load up intents
+intents = discord.Intents.all()
+
+bot = NerdlandBot(PREFIX, intents)
+
+# remove default help command
+bot.remove_command("help")
+
+# load event handlers
+bot.load_extension("nerdlandbot.eventhandlers.onmemberjoin")
+bot.load_extension("nerdlandbot.eventhandlers.onready")
+bot.load_extension("nerdlandbot.eventhandlers.oncommanderror")
+
+# load commands
+bot.load_extension("nerdlandbot.commands.notify")
+bot.load_extension("nerdlandbot.commands.help")
+bot.load_extension("nerdlandbot.commands.settings")
+bot.load_extension("nerdlandbot.commands.membercount")
+bot.load_extension("nerdlandbot.commands.random_user")
+bot.load_extension("nerdlandbot.commands.wombat")
+bot.load_extension("nerdlandbot.commands.youtube")
+bot.load_extension("nerdlandbot.commands.poll")
+
+# Initialize and start YouTube scheduler
+YOUTUBE_TOKEN = os.getenv("YOUTUBE_TOKEN")
+
+@bot.event
+async  def on_ready():
+    if YOUTUBE_TOKEN:
+        info("Starting YouTube scheduler")
+        check_and_post_latest_videos.start(bot)
+    else:
+        fatal("Not starting YouTube scheduler. Please provide a YOUTUBE_TOKEN in your .env file")
+
+bot.run(TOKEN)
