+++ nerdlandbot/bot/__init__.py
@@ -1,0 +1,1 @@
+from nerdlandbot.bot.nerdlandbot import NerdlandBot
