+++ nerdlandbot/eventhandlers/oncommanderror.py
@@ -1,9 +1,10 @@
 from typing import Any
 
 from discord.ext import commands
-from Translations.Translations import get_text as translate
-from Helpers.TranslationHelper import get_culture_from_context as culture
-from Helpers.log import error as log_error
+
+from nerdlandbot.translations.Translations import get_text as translate
+from nerdlandbot.helpers.TranslationHelper import get_culture_from_context as culture
+from nerdlandbot.helpers.log import error as log_error
 
 
 class OnCommandError(commands.Cog, name="on_command_error"):
