+++ nerdlandbot/helpers/emoji.py
