+++ nerdlandbot/persistence/abstractconfigstore.py
