+++ pyproject.toml
@@ -8,6 +8,8 @@ authors = ["Autom"]
 python = "^3.8"
 discord-py = "^1.5.1"
 python-dotenv = "^0.14.0"
+pandas = "^1.1.5"
+requests = "^2.25.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
